requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-cloud-firestore==2.11.1
google-cloud-storage==2.12.0
google-cloud-pubsub==2.18.0
python-dotenv==1.0.1
requests==2.31.0
flask_cors